requirements.txt

# Web Framework and Server
fastapi
uvicorn
uv
websockets

# LangChain and AI Model Integration
langchain-openai
# THIS IS THE CRUCIAL LINE
mcp-use[cli]

# Environment Variable Management
python-dotenv

# The Groq SDK is often a useful dependency
groq